add DisplayDiscardFilter

+++ SGPlayer/Classes/Core/SGCodec/SGVideoDecoder.h
@@ -16,8 +16,8 @@
 @property (nonatomic, assign) BOOL refcountedFrames;
 @property (nonatomic, assign) BOOL hardwareDecodeH264;
 @property (nonatomic, assign) BOOL hardwareDecodeH265;
-@property (nonatomic, copy) BOOL (^discardPacketFilter)(CMSampleTimingInfo timingInfo, int index, BOOL key);
-@property (nonatomic, copy) BOOL (^discardFrameFilter)(CMSampleTimingInfo timingInfo, int index);
+@property (nonatomic, copy) BOOL (^discardPacketFilter)(CMSampleTimingInfo timingInfo, NSUInteger index, BOOL key);
+@property (nonatomic, copy) BOOL (^discardFrameFilter)(CMSampleTimingInfo timingInfo, NSUInteger index);
 @property (nonatomic, assign) SGAVPixelFormat preferredPixelFormat;
 
 @end
